refactor to reduce duplicate code
